Доделал заодно старые правки
Проверка в Attacker на перезарядку
IsAlive теперь свойство Health
Flip больше не делает проверку дважды
Метод проверки Положительности входящего урона убран e Health
TargetFinder тоже чуть переделан

Не понял только про
Died?.Invoke(); - не надо дергать событие больше одного раза. мы вроде и так не дёрнем...но какая-то проверка чтоли нужна?

+++ Assets/Scripts/Character/Player/PickUpHandler.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class PickUpHandler: MonoBehaviour
 {
     public event Action CoinPicked;
-    public event Action<int> FirstAidKitPicked;
+    public event Action<float> FirstAidKitPicked;
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
